0.29.0 (#2)
* 0.29.0

* update apt

* make X_C_Order smaller

* code comments

* OrderGetters

* cleanup

+++ src/main/java/org/compiere/order/X_C_OrderPaySchedule.java
@@ -61,7 +61,7 @@ public class X_C_OrderPaySchedule extends PO implements I_C_OrderPaySchedule {
      * @return Order
      */
     public int getOrderId() {
-        Integer ii = (Integer) getValue(COLUMNNAME_C_Order_ID);
+        Integer ii = getValue(COLUMNNAME_C_Order_ID);
         if (ii == null) return 0;
         return ii;
     }
@@ -92,7 +92,7 @@ public class X_C_OrderPaySchedule extends PO implements I_C_OrderPaySchedule {
      * @return Calculated amount of discount
      */
     public BigDecimal getDiscountAmt() {
-        BigDecimal bd = (BigDecimal) getValue(COLUMNNAME_DiscountAmt);
+        BigDecimal bd = getValue(COLUMNNAME_DiscountAmt);
         if (bd == null) return Env.ZERO;
         return bd;
     }
@@ -130,7 +130,7 @@ public class X_C_OrderPaySchedule extends PO implements I_C_OrderPaySchedule {
      * @return Amount of the payment due
      */
     public BigDecimal getDueAmt() {
-        BigDecimal bd = (BigDecimal) getValue(COLUMNNAME_DueAmt);
+        BigDecimal bd = getValue(COLUMNNAME_DueAmt);
         if (bd == null) return Env.ZERO;
         return bd;
     }
