+++ src/main/java/org/compiere/order/X_M_ShipperLabels.java
@@ -63,7 +63,7 @@ public class X_M_ShipperLabels extends BasePOName implements I_M_ShipperLabels {
      * @return Shipper Labels Configuration
      */
     public int getShipperLabelsCfgId() {
-        Integer ii = (Integer) getValue(COLUMNNAME_M_ShipperLabelsCfg_ID);
+        Integer ii = getValue(COLUMNNAME_M_ShipperLabelsCfg_ID);
         if (ii == null) return 0;
         return ii;
     }
